a swatch changes shades based on whether a numeric threshold is crossed
See numbers.html, which is based on a jQuery example of sliders that
affect the color of a swatch. Here, I demonstrate that the sliders can
pass values to a numerical calculation, and that the color of the swatch
can change based on whether the number clears a certain threshold. Next
step, I think, is to write a function that reads in a list of numbers
and outputs n quantile cutpoints.